Fixed Start Time Offset
Implementation has been modified to utilize the firstplay event to set
the custom start time offset of the video.

If in the data-setup array an attribute of "starttime" is specified,
then the video will start playback from the passed index in seconds.

Note: Video will only start from this index on first play only.